split into c/c++ for module mixed non- and designated-initializers
